refactor(web): move build-time branding beside its siblings
No behaviour change. Tests are unchanged bar an import reorder, and dev and
prod builds emit the same document, robots.txt and sitemap.xml as before.

Move Function takes environment-branding out of src/lib. It never reaches the
bundle — only vite.config imports it — and site-files, which does the same kind
of build-time generation, already lived in scripts. The two now sit together,
with src/lib/environments left holding the facts both the app and the build
need.

That move makes the duplicate honest: site-files declared PUBLIC_ORIGIN and
environment-branding declared PRODUCTION_ORIGIN, the same literal for the same
reason, which the merge with #1254 introduced. One constant lives in
environments now and both read it, so isPublicOrigin keeps its name and its
meaning without keeping its own copy. site-files.test.ts still states the
literal itself, deliberately: a test that imports the value it exercises stops
being an independent check.

Extract Variable turns the two near-identical mark images into a table of the
one thing that differs between them.
