+++ apps/web/scripts/test-globals.d.ts
